requirements.txt
Pillow
keyboard
pyautogui
pygetwindow
pywin32
psutil
pyinstaller